@@ -5,9 +5,9 @@ class RaterController < ApplicationController
       obj = params[:klass].classify.constantize.find(params[:id])
       obj.rate params[:score].to_i, current_user, params[:dimension]
 
-      render :json => true
+      render json: true
     else
-      render :json => false
+      render json: false
     end
   end
 
