@@ -1,4 +1,4 @@
-class Comment < ActiveRecord::Base
+class Comment < ApplicationRecord
   belongs_to :commentable, :polymorphic => true
   has_many :comments, :as => :commentable
   belongs_to :user
@@ -7,5 +7,4 @@ class Comment < ActiveRecord::Base
     return @show if defined?(@show)
     @post = commentable.is_a?(Show) ? commentable : commentable.show
   end
-
 end
