@@ -1,4 +1,4 @@
-class Episode < ActiveRecord::Base
+class Episode < ApplicationRecord
   belongs_to :show
   has_many :comments, :as => :commentable
 end
