@@ -1,4 +1,4 @@
-class Rate < ActiveRecord::Base
+class Rate < ApplicationRecord
   belongs_to :rater, :class_name => "User"
   belongs_to :rateable, :polymorphic => true
 
