@@ -1,4 +1,4 @@
-class ShowsGenre < ActiveRecord::Base
+class ShowsGenre < ApplicationRecord
   belongs_to :show
   belongs_to :genre
 end
