@@ -1,4 +1,4 @@
-class UsersEpisode < ActiveRecord::Base
+class UsersEpisode < ApplicationRecord
   belongs_to :user
   belongs_to :episode
 end
