@@ -1,4 +1,4 @@
-class CreateForeignKeys < ActiveRecord::Migration
+class CreateForeignKeys < ActiveRecord::Migration[4.2]
   def change
     add_foreign_key :users_shows, :shows
     add_foreign_key :users_shows, :users
