@@ -1,4 +1,4 @@
-class CreateGenres < ActiveRecord::Migration
+class CreateGenres < ActiveRecord::Migration[4.2]
   def change
     create_table :genres do |t|
       t.string :name,     null:false
