@@ -1,4 +1,4 @@
-class CreateUsersEpisodes < ActiveRecord::Migration
+class CreateUsersEpisodes < ActiveRecord::Migration[4.2]
   def change
     create_table :users_episodes do |t|
       t.references :episode, show: true
